pyproject.toml
# Copyright 2023 Yaroshenko Dmytro (https://github.com/o-murphy)

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "pydfuutil"
version = "0.11.0b0"
authors = [
    { name="o-murphy", email="thehelixpg@gmail.com" },
]
description = "PyDfuUtil - Pure python fork of dfu-util wrappers to libusb"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["dfu_util", "dfu-util", "pydfu", "libusb", "python", "python3"]
license = {file = "LICENSE"}
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = ['pyusb', 'libusb-package']


[project.urls]
"Homepage" = "https://github.com/o-murphy/pydfuutil"
"Bug Reports" = "https://github.com/o-murphy/pydfuutil/issues"
"Source" = "https://github.com/o-murphy/pydfuutil"


[tool.setuptools]
py-modules = ["pydfuutil"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pydfuutil*"]


[tool.setuptools.dynamic]
version = {attr = "pydfuutil.__version__"}


[project.optional-dependencies]
build = [
    "build"
]
rich = [
    "rich"
]
test = [
    "pytest",
    "pylint",
]

[project.scripts]
pydfuutil = "pydfuutil.__main__:main"
pydfuutil-suffix = "pydfuutil.suffix:main"
pydfuutil-prefix = "pydfuutil.prefix:main"
pydfuutil-lsusb = "pydfuutil.lsusb:main"
